refactor(EditableAnswer): 답변이 있을 때 중첩되어 분기 되는 부분 컴포넌트로 분리
